fixed lab 4 x10

+++ tsareva.yana/I3/golden_ratio.cpp
@@ -5,7 +5,7 @@
 
 size_t goldenRatio(size_t capacity)
 {
-  const size_t gRatioK = ((1 + std::sqrt(5)) / 2);
+  const double gRatioK = ((1 + std::sqrt(5)) / 2);
   size_t max_size_t = std::numeric_limits< size_t >::max();
   if ((max_size_t - ((gRatioK * capacity) - capacity)) <= capacity)
   {
